Fixed a drag and drop issue
Fixed a drag and drop issue in the engine when the container of blocks is loaded with a lot of blocks. In that case, when we try to drag and drop an element from the bottom of the list it disappears. 
To reproduce this bug, just fill the container of blocks with a lot of blocks and scroll down to the bottom and try to drag and drop the last one for example.